pyproject.toml
[project]
name = "endfield-sarkaz"
version = "0.1.0"
description = "Sarkaz end-to-end translation research toolkit"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  "accelerate>=0.34.2",
  "datasets>=3.0.0",
  "peft>=0.13.0",
  "pyahocorasick>=2.1.0",
  "sentencepiece>=0.2.0",
  "tiktoken>=0.8.0",
  "tokenizers>=0.20.0",
  "torch>=2.4.0",
  "transformers>=4.45.0"
]

[project.optional-dependencies]
train = [
  "trl>=0.11.1"
]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.6.7"
]

[project.scripts]
skz-generate-data = "training.data_generator:main"
skz-train-base = "training.base_train:main"
skz-train-tokenizer = "training.tokenizer_train:main"
skz-merge-tokenizer = "training.merge_tokenizer:main"
skz-build-trie = "inference.trie_builder:main"
skz-decode = "inference.sarkaz_decoder:main"
skz-eval = "scripts.evaluate:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling>=1.26.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["training", "inference", "scripts"]

[tool.hatch.build.targets.sdist]
include = [
  "training/**",
  "inference/**",
  "scripts/**",
  "README.md",
  "pyproject.toml"
]
